requirements.txt
Django==4.2.28
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
python-dotenv==1.2.1
pytest==8.4.2
pytest-django==4.11.1
pytest-cov==7.0.0
flake8==7.3.0
